Numeric Data Types

+++ myfirstprogram.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main() {
     // Creating variables
-    float myNum = 1000.75;
+    double myNum = 1000.75;
     cout << myNum << endl;
 
     return 0;
